Create pong1UserSwitchRotary
In this version of the code, we've added a rotary switch as an input device to control the movement of the paddle. The switch is connected to GPIO pin 7 and reads its state using the Pico's built-in ADC module. We set up a new ADC channel using the machine.ADC() class, and then read the value of the channel using the adc_c.value() method.

We then use the value read from the rotary switch to control the movement of the paddle. If the switch position is greater than 800, we move the paddle up; if it's less than 200, we move the paddle down. We also keep the previous button controls for moving the paddle, so that the player can choose to use either input device.

With these modifications, your MicroPython Pong game can now be played using both buttons and a rotary switch as input devices for controlling the paddle.