Create Pong1User1Encoder
In this version of the code, we've replaced the rotary switch with a rotary encoder as the input device to control the movement of the paddle. The encoder is connected to GPIO pins 6 and 7, and we use the machine.RotaryEncoder() class to set it up.

We then read the value of the encoder using the encoder.value() method, which returns the number of steps the encoder has moved since the last time it was reset. We check if the value is greater than zero, in which case we move the paddle up, or less than zero, in which case we move the paddle down. We also reset the encoder using the encoder.reset() method after each movement, so that the encoder value doesn't keep increasing indefinitely.